+++ models/src/main/kotlin/motif/models/Type.kt
@@ -15,10 +15,8 @@
  */
 package motif.models
 
-import motif.ast.IrAnnotation
-import motif.ast.IrMethod
-import motif.ast.IrParameter
-import motif.ast.IrType
+import motif.ast.*
+import javax.inject.Qualifier
 
 /**
  * This is the primitive type representation that Motif operates on. It's a data class so .equals() is valid for
@@ -53,6 +51,27 @@ data class Type(val type: IrType, val qualifier: IrAnnotation?) : Comparable<Typ
 
         fun fromReturnType(method: IrMethod): Type {
             return Type(method.returnType, method.getQualifier())
+        }
+
+        private fun IrAnnotated.getQualifier(): IrAnnotation? {
+            val qualifier = annotations.find { annotation ->
+                val annotationClass: IrClass = annotation.type.resolveClass() ?: return@find false
+                annotationClass.hasAnnotation(Qualifier::class)
+            } ?: return null
+
+            val members = qualifier.members
+            if (members.size > 1) {
+                throw InvalidQualifier(this, qualifier.type)
+            }
+
+            if (members.size == 1) {
+                val member = members[0]
+                if (member.name != "value" || member.returnType.qualifiedName != "java.lang.String") {
+                    throw InvalidQualifier(this, qualifier.type)
+                }
+            }
+
+            return qualifier
         }
     }
 }