+++ tests/src/main/java/testcases/E052_custom_qualifier_with_nonstring_value/Scope.java
@@ -13,30 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package testcases.T006_custom_qualifier_with_fields;
+package testcases.E052_custom_qualifier_with_nonstring_value;
 
 @motif.Scope
 public interface Scope {
 
-    String string();
-
     @motif.Objects
     class Objects {
 
-        String s(
-                @CustomQualifier(first = "1", second = "2") String a,
-                @CustomQualifier(first = "3", second = "4") String b) {
-            return a + b;
-        }
-
-        @CustomQualifier(first = "1", second = "2")
+        @CustomQualifier(1)
         String a() {
             return "a";
-        }
-
-        @CustomQualifier(first = "3", second = "4")
-        String b() {
-            return "b";
         }
     }